migrate biometrics link to typescript
