+++ src/Settings/BiometricsLink/BiometricsLinkScreen.stories.tsx
@@ -1,8 +1,6 @@
-// @flow
-
 import {storiesOf} from '@storybook/react-native'
 import React from 'react'
 
-import BiometricsLinkScreen from './BiometricsLinkScreen'
+import {BiometricsLinkScreen} from './BiometricsLinkScreen'
 
 storiesOf('BiometricsLinkScreen', module).add('Default', () => <BiometricsLinkScreen />)
