+++ src/Settings/BiometricsLink/index.ts
@@ -1,0 +1,1 @@
+export * from './BiometricsLinkScreen'
